AEAD: KMAC/cSHAKE - set auth key to 256 bits
By lowering the size of the auth key, the initialization of the auth
cSHAKE state requires one KECCAK operation less than before. As 256 bits
are considered to be sufficient considering the security strength of the
algorithm to be 256 bits, the size reduction has no cryptographic
impact.

This change implies a modification of the test vectors.

Also, add another test to compare the output of the algorithm with a
plain cSHAKE / KMAC operation to validate that the algorithm works
compliant to the design.

Signed-off-by: Stephan Mueller <smueller@chronox.de>