cSHAKE DRNG: conform better to SP800-185
The cSHAKE specification defines that N is the function name. Thus, the
cSHAKE DRNG uses N as a function name and uses S as the key. This
applies to both, the seeding as well as the generation operation.

Technically this implies a simple switch in arguments to the cSHAKE
init function.

Suggested-by: Joachim Vandersmissen
Signed-off-by: Stephan Mueller <smueller@chronox.de>