cSHAKE AEAD: conform better to SP800-185
The cSHAKE specification defines that N is the function name. Thus, the
cSHAKE DRNG uses N as a function name and uses S as the key. This
applies to both, the main crypto-cSHAKE instance as well as the
authentication cSHAKE instance. The IV is used with the X parameter.

Technically this implies a simple switch in arguments to the cSHAKE
init function as well as a new Keccak absorb invocation for the IV.

Suggested-by: Joachim Vandersmissen
Signed-off-by: Stephan Mueller <smueller@chronox.de>