simplify kyber_enc implementation
The reference implementation hashes the output of the RNG which is not
required for appropriate RNGs. As leancrypto only uses appropriate RNGs,
the hashing step is now removed.

To still allow comparison of the test results with the reference
implementation, add the hashing step to the test RNG that is used for
the kyber_enc invocation.

The change implies that the KyberIES test result is changed, too.

Signed-off-by: Stephan Mueller <smueller@chronox.de>