use link_whole
Up to now, the static libraries wrapping the AVX code were only linked in
by accident. This lead to the issue that the dilithium library was not
linked in at all.

Using link_whole, all libraries are now linked in by force. This implies
that some symbols with identical names have to be renamed.

Signed-off-by: Stephan Mueller <smueller@chronox.de>