Kyber KEX: allow caller-provided data to be added to KDF
A user may provide nonce data or other information that is added
to the KDF to obtain the shared secret. This shall allow Kyber to
be used consistently with contemporary protocols. When using
NULL as user-provided data, the Kyber KEX shared secret is
identical to the original Kyber KEX code.

Signed-off-by: Stephan Mueller <smueller@chronox.de>