Kyber KDF: remove SHA3-256 hash
The round 3 of the Kyber specification contains the KDF deriving the final
shared secret as:

	SS <- SS <- SHAKE256(Kyber-SS || SHA3-256(Kyber-CT), 32)

Considering that the Kyber-CT is inserted into a SHAKE256 anyway,
calculating a preceding hash is mathematically and cryptographically
superflouous. This, this patch changes the KDF to

	SS <- SHAKE256(Kyber-SS || Kyber-CT, 32)

The application of the KDF is still adjusted to allow the caller to specify
the amount of bytes he wants to obtain from the KDF. I.e. the final KDF
definition therefore is:

	SS <- SHAKE256(Kyber-SS || Kyber-CT, requested SS length)

Signed-off-by: Stephan Mueller <smueller@chronox.de>