Dilithium: increase size of ctilde to lambda * 2
According to FIPS 204, the ctilde value has to be lambda * 2 in size.
This implies that for level 3 and level 5, ctilde is 384 and 512 bits,
respectively. This also implies that the signature is increased by 16
and 32 bytes, respectively.

The change is verified (a) against CRYSTALS implementation to still
calculate the same signature (except for the new bytes), and (b) against
NIST's reference implementation.

Signed-off-by: Stephan Mueller <smueller@chronox.de>