Dilithium: skip (un)packing of c~ from/to signature
The c~ value in the signature is a simple bit-stream which is used as is.
Thus, a pack/unpack so far consisted of a simple copy of the data into
a separate buffer. This is a waste of CPU cycles and memory. Thus, the
c~ value is immediately stored into the signature buffer during siggen
and directly used from the signature buffer during sigver.

Signed-off-by: Stephan Mueller <smueller@chronox.de>