Ascon: extract Keccak part out
This is a preparation step to add the Ascon permutation and allow both to exist jointly.

Although no user-visible changes are present, the memory structure used for Ascon is
updated slightly to use a pointer to a memory buffer. That buffer is allocated by
either Ascom-Keccak or Ascom native allocation functions.

Signed-off-by: Stephan Mueller <smueller@chronox.de>