Ascon-Keccak: use only byte values in IV construction
As the IV is of limited size, use only byte values in the IV to reference
different algorithm properties. Each value cannot technically be larger
than 256 which implies that an 8-bit value suffices.

The change does not alter the algorithmic logic, but since the representation
of the IV is different, the resulting ciphertext and tags are different.

Signed-off-by: Stephan Mueller <smueller@chronox.de>