Kyber: Add API automatically selecting Kyber type
The old Kyber 1024 API calls of lc_kyber_* are moved to lc_kyber_1024_*.
As a replacement, a generic lc_kyber_* API is added that is only
living in the lc_kyber.h header file which selects the right
Kyber type API depending on the input parameters. The Kyber type is
stored in the data structures for the Kyber keys, ciphertext and
shared secrets to limit the user need to track the Kyber types.

Signed-off-by: Stephan Mueller <smueller@chronox.de>