remove support for AES no SBOX
This implementation is horribly slow (50s for the S-Box C implementation
for 1GB vs 75 minutes for the non-S-Box implementation). Besides, the side
channel issue is not truly removed as there is a branch on the private
key data in gf_mulinv (see the unpoison calls there).

Finally, the implementation is not yet working on big-endian systems.

Signed-off-by: Stephan Mueller <smueller@chronox.de>