ML-DSA: support retaining AHat
FIPS 204 allows callers to retain the internal value of AHat which is
the expanded RHO value present in the secret key and public key.

Generally using AHat for a signature generation or verification
improves performance by about 15% and it reduces the amount of runtime
memory for the signature generation / verification by the size of AHat
(which is 256 * sizeof(uint32_t) * L * K) - which is quite substantial.

Signed-off-by: Stephan Mueller <smueller@chronox.de>