ASN.1 / X.509 / PKCS7 support
The addition includes an ASN.1 decoder which uses the message to be
decoded as input on one side and the decoding machine on the other
side. The decoding machine is defined by the different *asn1 files
which are convered into C/H files with the asn1_compiler application
part of this code tree.

The X.509 parser parses the message into an internal structure allowing
an immediate access to the data parts.

The X.509 policy checker offers various X.509 policies to be checked
against a parsed X.509 certificate.

The PKCS#7 parser parses message into an internal structure allowing
an immediate validation of the data. Yet, the validation must be
started with the PKCS#7 verifier.

Currently missing is a trust store and the associated checks. This
trust store will be added in the next round.