X.509: Fix DN encoding
That fixes the interoperability with Bouncy Castle. The solution was
to have only AVA in the RDN set, but then have multiple RDNs in the
sequence that forms RDNSequence. With this change, BC validates my
certificates with even 5 entries. Peter Guttman describes that in
his document (thanks for the link).

Signed-off-by: Stephan Mueller <smueller@chronox.de>