meson: enable to build only static library
Previously, there was no way to build only static libraries. Calling
`meson setup` with -Ddefault_library=static led to the following
error, because the static library is created multiple times with the
same name through the `static_library` and `library` functions:

  meson.build:633:18: ERROR: Tried to create target "leancrypto", but a target of that name already exists.

This patch makes it work by switching the call to `library` to
`shared_library`, taking into account of the "default_library" option
as well as the existing "efi" option.

Signed-off-by: Daiki Ueno <dueno@redhat.com>