x86: embed .note.gnu.property for Intel CET in assembly files
Instead of generating the note section with a linker option, this
embeds it in each .S file through a preprocessor macro. The main
motivation behind this is to simplify the build process when a shared
library is statically linking to leancrypto. As a bonus, this
mechanism could be used to support PACBTI on AArch64 in a future
extension.

Signed-off-by: Daiki Ueno <dueno@redhat.com>
Signed-off-by: Stephan Mueller <smueller@chronox.de>