ARMv8: enable PAC and BTI support
Using the description in [1], the LC_ASM_END macro is defined for
ARMv8 to enable PAC and BTI support:

$ readelf -n build/libleancrypto.so

Displaying notes found in: .note.gnu.property
  Owner                Data size        Description
  GNU                  0x00000010       NT_GNU_PROPERTY_TYPE_0
      Properties: AArch64 feature: BTI, PAC

The changes in the assembly code only ensures that the PAC
instructions are placed at the proper locations. Yet, the changes
do not alter the functionality.

This patch, however, disables the PAC/BTI support on Apple
because it yet would cause a SIGILL.

[1] https://community.arm.com/arm-community-blogs/b/architectures-and-processors-blog/posts/enabling-pac-and-bti-on-aarch64

Signed-off-by: Stephan Mueller <smueller@chronox.de>