HQC: add C implementation for 128/192/256
The code is based on the version provided by PQClean. The test vectors
are derived from the reference implementation KAT tools (the first 10
test vectors are used).

The PQClean implementation is changed such that it complies with the
reference implementation (see [1] for the reported issues). The changes
can also be identified with the TODO markers in the code.

[1] https://github.com/PQClean/PQClean/issues/593

Signed-off-by: Stephan Mueller <smueller@chronox.de>