rearchitecture of self-test
The current self test uses assert() when a self test fails. This kills the
library and the application. Although a self test should never occur, this
approach is an availability hazard.

Thus, this patch re-architects the self test by replacing the assert() calls
with a test manager that sets a bit for each passed / failed self test.
If a self test failed, its operation is simply blocked, but the rest of
the library remains functional. It is possible to trigger a new test
to perhaps see a passing result (again, should also not happen, but there
may be edge cases where such issue may occur).

Albeit all algorithms are unchanged as well as all self tests, the patch
touches all ciphers to change how the self test operates:

- the self test invocation as well as invocation check is moved to the
interface functions of an algorithm

- to aovid races as well as deadlocks, every self test bypasses the
interface check by having a *_nocheck function that invokes the gated
functionality directly without the self test check

- add to each regression test a verification that the self test execution
is properly recorded

- during startup of the library, all interfaces are gated until the
initialization function of lc_activate_library is called either via a
constructor or via lc_init - this avoids the accidental forgetting of
calling lc_init in case constructors are not present

- the change should reduce the runtime memory usage as the individual
`int tested` variables that only hold one bit are replaced by the
central variables that truly contain one bit per algorithm testing.

The whole patch yet are effectively be a noop to the cryptographic
algorithms.

Signed-off-by: Stephan Mueller <smueller@chronox.de>