Self-test manager: properly handle multithreaded environments
The self-test manager is rearchitected such that:

- only atomic variables hold the state

- each state transition is a fully atomic operation

This is achieved by using 3 bits for the state as defined by
`lc_alg_status_result` where each new state only adds a bit
to the state.

Furthermore, when the test state is "ongoing", add a tight busy
loop to wait on this flag as this state is transient and only very
short in nature. This busyloop ensures that only one self test is
triggered / performed during startup.

This change now fully and race-free supports multi-threaded
environments where two or more parallel threads initialize one
algorithm for the first time.

Signed-off-by: Stephan Mueller <smueller@chronox.de>