FIPS: rearchitect the generation of the FIPS control value
The build process now uses the host's objcopy to extract the ELF sections
of interest into a separate file, use a build_machine compiled version of
sha3-256sum to generate the digest of it and reinsert it into the
leancrypto-fips.so.

This now allows cross-compilation with FIPS integrity test support.

There is no functional change to leancrypto though.

Yet, this change now allows the removal of the special FIPS-ro ELF sections
throughout the library code to leave any static data at their natural section:
the .rodata section.

Signed-off-by: Stephan Mueller <smueller@chronox.de>