SHA2-256/512 asm: leave the K256/K512 table in the .text segment
The Cryptogam implementation leaves the K static data in the text segment.
Thus, leancrypto puts it there too. This helps also the FIPS integrity test
as the text segment is integrity-tested whereas rodata is not.

I just have the concern that the K data contains a gadget for adversaries
to misuse if they can apply a form of ROP or similar to execute code from
the K data.

Signed-off-by: Stephan Mueller <smueller@chronox.de>