FIPS: rearchitect service indicator and library initialization
The patch changes the library initialization in FIPS mode such that the
library does not exit any more on FIPS integrity error, but remains in
error state (all algorithms are gated). The status of the library will
reflect that.

This re-architecture also implies changes on how the library is built:

- Create a common static library implementing the full library with the
  exception of the FIPS integrity test or the noop thereof.

- Link the common static lib with the FIPS noop to get the final non-FIPS lib

- Link the common static lib with the FIPS integrity test to get the final
  FIPS lib

The change adds the algorithm type to the cipher handles to allow the caller
to use the cipher handle to obtain FIPS 140 status information about the
algorithm.

This change also covers:

- addition of key length check in FIPS mode to force at least 128 bits of
  key lengths. This necessitated the update of HMAC self tests as their
  keys were less than 128 bits.

- split up the Ascon-AEAD implementations to cover them with individual cipher
  handles

- Update rerun_selftest to use the library in FIPS mode

Signed-off-by: Stephan Mueller <smueller@chronox.de>