ML-DSA: Code reduction
The code changes move the loops around K and L into the main functions
allowing to remove the helper functions. This change is also intended
to add slight performnance increase by keeping the data multiple steps
operates on in cache.

Albeit of the larger change, it is "just" a move of code without
functional changes.

Signed-off-by: Stephan Mueller <smueller@chronox.de>