Linux kernel: allow leancrypto to be compiled in-tree statically bound
The change set allows a static compilation of leancrypto into kernel
binary making its services available from the start. This supports
out-of-the-box moduloe signing.

The changes to the sources only rename symbols to avoid clashes with the
vanilla kernel. This is required as the vanilla kernel has ASN.1 support
which is used in leancrypto as well. This implied the presence of some
duplicated symbols.

Signed-off-by: Stephan Mueller <smueller@chronox.de>