CTR DRBG: remove temporary buffer
To "manually" increment the counter value V, the code uses an 64-bit
integer operation. This was handled by using turning the value from
the big-endian buffer to an integer array using a temporary value.

This is changed by using a union and doing an inplace byteswap.

Furthermore, the change adds test code for using counter values that
are at most 8 bits. This code is commented out and only present for
testing and analyses, but not intended for production code.

Signed-off-by: Stephan Mueller <smueller@chronox.de>